Control system changes needed
Figured out how to use the PID on lbmotor. Need to change control scheme for speed and set up PID for all motors. 

We have been running at half speed this whole time. 

The gear ration from shaft to wheel is 1.75. Inverted that is our magic factor of 0.57!!! This needs to be fixed!